pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ankiflow-tts"
version = "0.1.0"
description = "CLI importer that generates Gemini TTS audio and creates Anki notes."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.12.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
ankiflow-tts = "ankiflow_tts.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
